+++ test/unit/fuzz-tools.test.js
@@ -302,6 +302,7 @@ import * as biRadsV322 from '../../lib/bi-rads-v322.js';
 import * as siewertV323 from '../../lib/siewert-v323.js';
 import * as wexnerV324 from '../../lib/wexner-v324.js';
 import * as lungRadsV325 from '../../lib/lung-rads-v325.js';
+import * as oRadsV326 from '../../lib/o-rads-v326.js';
 import * as tbTesting from '../../lib/tb-testing.js';
 
 const MODULES = {
@@ -549,6 +550,7 @@ const MODULES = {
   'siewert-v323.js': siewertV323,
   'wexner-v324.js': wexnerV324,
   'lung-rads-v325.js': lungRadsV325,
+  'o-rads-v326.js': oRadsV326,
   'tb-testing.js': tbTesting,
 };